+++ src/is-ios-safari.ts
@@ -1,4 +1,5 @@
 import isIos = require("./is-ios");
+import isIosFirefox = require("./is-ios-firefox");
 
 const webkitRegexp = /webkit/i;
 
@@ -9,5 +10,7 @@ function isWebkit(ua: string): boolean {
 export = function isIosSafari(ua?: string): boolean {
   ua = ua || window.navigator.userAgent;
 
-  return isIos(ua) && isWebkit(ua) && ua.indexOf("CriOS") === -1;
+  return (
+    isIos(ua) && isWebkit(ua) && ua.indexOf("CriOS") === -1 && !isIosFirefox(ua)
+  );
 };